+++ packages/miniprogram-element/src/component/slider.js
@@ -10,13 +10,13 @@ module.exports = {
     }, {
         name: 'max',
         get(domNode) {
-            const value = parseInt(domNode.getAttribute('max'), 10)
+            const value = parseFloat(domNode.getAttribute('max'), 10)
             return !isNaN(value) ? value : 100
         },
     }, {
         name: 'step',
         get(domNode) {
-            const value = parseInt(domNode.getAttribute('step'), 10)
+            const value = parseFloat(domNode.getAttribute('step'), 10)
             return !isNaN(value) ? value : 1
         },
     }, {
@@ -52,7 +52,7 @@ module.exports = {
     }, {
         name: 'blockSize',
         get(domNode) {
-            const value = parseInt(domNode.getAttribute('block-size'), 10)
+            const value = parseFloat(domNode.getAttribute('block-size'), 10)
             return !isNaN(value) ? value : 28
         },
     }, {
